requirements.txt
alias_free_torch
diffusers
einops
librosa
pandas
PyYAML
soundfile
tqdm
transformers
vector_quantize_pytorch
torchaudio
julius
torch_stoi
flatten-dict
tensorboard
randomname
argbind
sentencepiece
accelerate